Revert "Update more packets"
This reverts commit e8c3e98cdaa48eee76ba1c7e0baa920d957d97f9.

+++ src/main/java/cn/nukkit/network/protocol/PlayerAuthInputPacket.java
@@ -43,8 +43,8 @@ public class PlayerAuthInputPacket extends DataPacket {
 		this.moveZ = this.getFloat();
 		this.headYaw = this.getFloat();
 		this.inputFlags = this.getUnsignedVarLong();
-		this.inputMode = (int) this.getUnsignedVarInt();
-		this.playMode = (int) this.getUnsignedVarInt();
+		this.inputMode = this.getUnsignedVarInt();
+		this.playMode = this.getUnsignedVarInt();
 		if (this.playMode == VR) {
 			this.vrGazeDirection = this.getVector3();
 		}
